Restructuring tests and small changes to code after review
