+++ pyaerocom/colocation_auto.py
@@ -203,7 +203,7 @@ class ColocationSetup(BrowseDict):
                  obs_name=None, model_name=None,
                  save_coldata=True,
                  gridded_reader_id='ReadGridded',
-                 data_id=None, **kwargs):
+                 **kwargs):
 
         if isinstance(obs_vars, str):
             obs_vars = [obs_vars]
@@ -226,7 +226,6 @@ class ColocationSetup(BrowseDict):
         self.read_opts_ungridded = read_opts_ungridded
         self.obs_ts_type_read = obs_ts_type_read
         self.gridded_reader_id = gridded_reader_id
-        self.data_id = data_id
         self.model_use_vars = model_use_vars
         self.model_add_vars = model_add_vars
         self.model_keep_outliers = model_keep_outliers
